feat: alarms 댓글 및 대댓글 작성시 알람 보내기, 댓글 및 대댓글 수정삭제시 작성자만가능하게하기2

+++ services/graphql/src/apis/replies/reply.resolver.ts
@@ -34,8 +34,9 @@ export class ReplysResolver {
     @Args('updateReplyInput') updateReplyInput: UpdateReplyInput,
     @Context() context: IContext, //
   ): Promise<Reply> {
+    const userId = context.req.user.id
     return this.replysService.update(
-      JSON.parse(JSON.stringify({ updateReplyInput })),
+      JSON.parse(JSON.stringify({ updateReplyInput, userId })),
     );
   }
 
@@ -45,6 +46,7 @@ export class ReplysResolver {
     @Args('replyId') replyId: string, //
     @Context() context: IContext, //
   ): Promise<string> {
-    return this.replysService.delete({ replyId });
+    const userId = context.req.user.id
+    return this.replysService.delete({ replyId, userId });
   }
 }
